Refactor(fe): `InviteCode`, `Menu` 상태 관리 로직을 페이지에서 `MyReservation` 컴포넌트 내부로 이동
